% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_analysis.R
\name{bca_analysis}
\alias{bca_analysis}
\title{Protein quantification from BCA}
\usage{
bca_analysis(standards_df, unknowns_df)
}
\arguments{
\item{standards_df}{Dataframe with protein standard concentrations, ug/ml, as column names and raw absorbance as values. Standard control column name must be 0. Required check.names = False.}

\item{unknowns_df}{Dataframe with unique ID as column names and raw absorbance as values. Required check.names = False.}
}
\value{
Dataframe of BCA results
}
\description{
Protein quantification from BCA
}
