% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix_example.R,
%   R/example_matrix_data.R
\name{create_matrix_example}
\alias{create_matrix_example}
\title{Create Matrix Example}
\usage{
create_matrix_example(
  replicate_n = 4,
  feature_n = 1000,
  seed = 123,
  a = "Treated",
  b = "Control"
)

create_matrix_example(
  replicate_n = 4,
  feature_n = 1000,
  seed = 123,
  a = "Treated",
  b = "Control"
)
}
\arguments{
\item{replicate_n}{Number of replicates (default 4)}

\item{feature_n}{Number of features (default 1000)}

\item{seed}{Seed for the random number generator (default 123)}

\item{a}{Column label prefix for group a (default "Treated")}

\item{b}{Column label prefix for group b (default "Control")}
}
\value{
A transposed data frame with sample names as row names and feature names as column names

A matrix with \code{replicate_n} columns and \code{feature_n} rows
}
\description{
Generate Example Expression Data for PCA Plotting
}
\examples{
create_matrix_example()
create_matrix_example(replicate_n = 5, feature_n = 80, seed = 123, a = "Experiment1", b = "Experiment2")
}
