% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proteomic_example.R
\name{create_proteomic_example}
\alias{create_proteomic_example}
\title{Random protemic data for examples}
\usage{
create_proteomic_example(seed = 1, length = 1000)
}
\arguments{
\item{seed}{Sets the seed}

\item{length}{Length of the data frame}
}
\value{
Dataframe of Log2 fold change and Log10 p-values
}
\description{
Random protemic data for examples
}
\examples{
create_proteomic_example(2, 10)
create_proteomic_example(seed = 2, length = 10)
}
