% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dep_1.R
\docType{data}
\name{example_dep_1}
\alias{example_dep_1}
\title{Example proteomic differential expression data}
\format{
An object of class \code{data.frame} with 1356 rows and 20 columns.
}
\source{
DEP package label free quantification proteomic data called UbiLength and originating from Zhang et al. 2017. Differential expression analysis was performed comparing condition Ubiquitin 4 and control
}
\usage{
example_dep_1
}
\description{
Example proteomic differential expression data
}
\examples{
data(example_dep_1)
}
\keyword{datasets}
