% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_proteomic_data}
\alias{example_proteomic_data}
\title{Example data to use when trying out function examples}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
example_proteomic_data
}
\description{
Log2 fold change and Log10 pvalue data for 1000 observations. Created using create_proteomic_example.R
\describe{
\item{One}{Log2 fold change}
\item{Two}{Log10 p-value}
}
}
\keyword{datasets}
