% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_peaks.R
\name{format_peaks}
\alias{format_peaks}
\title{Format Peaks}
\usage{
format_peaks(
  file,
  a,
  b,
  columns.a,
  columns.b,
  accession.split = c("sp", "Accession", "Name")
)
}
\arguments{
\item{file}{Character string with the path to the proteins.csv.}

\item{a}{A Character string representing the name of the first group of replicates.}

\item{b}{Character string representing the name of the second group of replicates.}

\item{columns.a}{Integer Vector with the indices of columns to be assigned the prefix a}

\item{columns.b}{Integer Vector with the indices of columns to be assigned the prefix b}

\item{accession.split}{Character Vector with the names of the separated accession columns. Defaults to c("sp", "Accession", "Name").}
}
\value{
dataframe
}
\description{
A function to format peaks proteins.csv file. The function reads a file, adds new columns with the names of the groups,
separates an accession column to get the accession ID on its own, and extracts gene names from a description
column.
}
\examples{
format_peaks("./proteins.csv", "Test", "Control", 15:17, 18:20, c("sp", "Accession", "Name"))
}
