% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_kegg.R
\name{gsea_kegg}
\alias{gsea_kegg}
\title{gsea_kegg}
\usage{
gsea_kegg(df, entrezID_column, s2n_column, pvalue_cutoff = 1)
}
\arguments{
\item{df}{Dfferential expression dataframe}

\item{entrezID_column}{Character string for the name of the column containing the Entrez IDs}

\item{s2n_column}{Character string for the name of the column containing the signal to noise data}

\item{pvalue_cutoff}{Numeric value for the p value cutoff used in the GSEA KEGG results}
}
\value{
Dataframe of GSEA KEGG results
}
\description{
Function to create a gene list and perform GSEA using the KEGG database
}
