% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_plot.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Missing data plot}
\usage{
missing_data_plot(
  df,
  sample,
  name,
  value,
  group,
  x_label = "Samples",
  y_label = "Number of proteins",
  title = NULL,
  tag = NULL,
  legend_label = NULL
)
}
\arguments{
\item{legend_label}{}
}
\value{
plot
}
\description{
Missing data plot
}
