% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot PCA Results}
\usage{
pca_plot(
  matrix,
  center = T,
  scale = T,
  a = "Treated",
  b = "Control",
  a_replicates,
  b_replicates,
  a_colour = "#D55E00",
  b_colour = "#3DB7E9",
  text_size = 20,
  title = NULL,
  tag = NULL,
  legend_label = NULL
)
}
\arguments{
\item{center}{Logical; if \code{TRUE}, variables will be centered (mean-subtracted) prior to PCA.}

\item{scale}{Logical; if \code{TRUE}, variables will be scaled (divided by standard deviation) prior to PCA.}

\item{a}{Character string representing the name of the first group of replicates.}

\item{b}{Character string representing the name of the second group of replicates.}

\item{a_replicates}{Integer value representing the number of replicates in the first group.}

\item{b_replicates}{Integer value representing the number of replicates in the second group.}

\item{a_colour}{Hexadecimal color code for the first group.}

\item{b_colour}{Hexadecimal color code for the second group.}

\item{text_size}{Integer value specifying the text size in the plot.}

\item{title}{Character string for the plot title.}

\item{tag}{Character string for the plot tag.}

\item{legend_label}{Character string for the plot legend label.}

\item{df}{Matrix containing IDs (accession or gene) as columns and samples as rows.}
}
\value{
A scatter plot of PCA results highlighting the two groups of replicates.
}
\description{
This function creates a scatter plot of principal component analysis (PCA) results, highlighting two groups of replicates.
}
\examples{
pca_plot(matrix = example_matrix_data, a_replicates = 4, b_replicates = 4)

}
