% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_plot.R
\name{point_plot}
\alias{point_plot}
\title{Point plot}
\usage{
point_plot(
  df,
  x,
  y,
  col = "condition",
  sd,
  formula = y ~ stats::poly(x, 2),
  line_size = 0.7,
  error_bar_factor = 0.02,
  expand_y_factor = 0.2,
  expand_x_factor = 0.1,
  legend_position = c(0.8, 0.2),
  x_label = NULL,
  y_label = NULL,
  col_label = NULL,
  title = NULL,
  tag = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{x}{A character string giving the x values}

\item{y}{A character string giving the y values}

\item{col}{A character string giving the colour values; defaults to "condition"}

\item{sd}{A character string giving the standard deviation values}

\item{formula}{A formula for the curve; defaults to y~stats::poly(x,2)}

\item{line_size}{A value for the line thickness; defaults to 0.7}

\item{error_bar_factor}{A value that is multiplied by the x axis range to set the error bar width; defaults to 0.02}

\item{expand_y_factor}{A values to increase the y axis by; defaults to 0.2}

\item{expand_x_factor}{A values to increase the x axis by; defaults to 0.1}

\item{legend_position}{A vector for the legend position; defaults to c(0.8, 0.2)}

\item{x_label}{A character string giving the x axis label}

\item{y_label}{A character string giving the y axis label}

\item{col_label}{A character string giving the colour label}

\item{title}{A character string giving the title}

\item{tag}{A character string giving the tag}
}
\value{
Plot
}
\description{
Plot data with standard deviation and a curve
}
