% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2n.R
\name{s2n}
\alias{s2n}
\title{Signal to Noise Ratio}
\usage{
s2n(x1, x2)
}
\arguments{
\item{x1}{A numeric vector of values.}

\item{x2}{A numeric vector of values.}
}
\value{
A scalar value representing the signal to noise ratio of the input vectors.
}
\description{
This function calculates the absolute signal to noise ratio of two input vectors.
}
\examples{
x1 <- c(1, 2, 3, 4)
x2 <- c(4, 3, 2, 1)
s2n(x1, x2)
}
