% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_plot.R
\name{viability_plot}
\alias{viability_plot}
\title{Viability plot}
\usage{
viability_plot(
  df,
  x = "conc",
  y = "mean",
  group = "condition",
  sd = "sd",
  size_line = 1,
  size_point = 2,
  size_text = 20,
  width_error_bar = 0.05,
  position_legend = c(0.2, 0.2),
  label_x = NULL,
  label_y = "Cell Viability (\%)",
  label_legend = NULL,
  break_labels_x = NULL,
  breaks_x = NULL,
  breaks_y = NULL,
  limit_x = NULL,
  limit_y = c(0, 120),
  title = NULL,
  tag = NULL,
  transparent = T,
  colours = c("#2271B2", "#3BB9FF", "#D55E00")
)
}
\arguments{
\item{df}{Dataframe}

\item{x}{A character string giving the x values; defaults to "conc"}

\item{y}{A character string giving the y values; defaults to "mean"}

\item{group}{A character string giving the group names; defaults to "condition"}

\item{sd}{A character string giving the standard deviation values; defaults to "sd"}

\item{size_line}{A value for the line thickness; defaults to 1}

\item{size_point}{A value for the point size; defaults to 2}

\item{size_text}{A value for the text size; defaults to 20}

\item{width_error_bar}{A value for the error bar width; defaults to 0.05}

\item{position_legend}{A vector for the legend position; defaults to c(0.2, 0.2)}

\item{label_x}{A character string giving the x axis label}

\item{label_y}{A character string giving the y axis label; defaults to "Cell Viability (\%)"}

\item{label_legend}{A character string giving the legend label}

\item{break_labels_x}{A character string giving the x axis break labels. Same length as breaks}

\item{breaks_x}{A vector giving the x axis breaks}

\item{breaks_y}{A vector giving the y axis breaks}

\item{limit_x}{A vector giving the x axis limits}

\item{limit_y}{A vector giving the y axis limits; defaults to c(0,120)}

\item{title}{A character string giving the title}

\item{tag}{A character string giving the tag}

\item{transparent}{A logical value for a transparent background; defaults to True}

\item{colours}{A character vector giving the colours to use; defaults to c("#2271B2", "#3BB9FF","#D55E00")}
}
\value{
Plot
}
\description{
Plot viability data with standard deviation and a curve
}
