% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot for proteomic data}
\usage{
volcano_plot(
  df,
  x = "Log2fold",
  y = "Log10pvalue",
  p_value = 0.05,
  log2foldvalue = 1,
  point_size = 1,
  text_size = 20,
  axis_label_x = "Fold change",
  axis_label_y = bquote(italic("P") * " value"),
  transparent = T
)
}
\arguments{
\item{df}{Dataframe containing Log2 fold change and Log10 p value columns}

\item{x}{A character string giving the Log2 fold change column name}

\item{y}{A character string giving the Log10 p value column name}

\item{p_value}{p value cut off; defaults to 0.05}

\item{point_size}{Point size; defaults to 1}

\item{text_size}{Text size; defaults to 20}

\item{axis_label_x}{X axis label; defaults to "Fold change"}

\item{axis_label_y}{Y axis label; defaults to bquote(italic("P")*" value")}

\item{transparent}{A logical value for a transparent background; defaults to True}

\item{Log2foldvalue}{Log2 fold change cut off; defaults to 1}
}
\value{
Volcano plot
}
\description{
Volcano plot for proteomic data
}
\examples{
volcano_plot(example_proteomic_data)
}
