% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Coverage Plot}
\usage{
coverage_plot(
  df,
  number_of_samples,
  percentage,
  tally,
  a,
  b,
  group,
  a_colour = "#F748A5",
  b_colour = "#3BB9FF",
  x_label = "Number of Samples",
  y_label = "Percentage of Proteins",
  title = NULL,
  tag = NULL,
  legend_label = NULL
)
}
\arguments{
\item{legend_label}{}
}
\value{
Plot
}
\description{
Coverage Plot
}
