% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Coverage Plot}
\usage{
coverage_plot(
  df,
  number_of_samples,
  percentage,
  tally,
  a,
  b,
  group,
  a_colour = "#F748A5",
  b_colour = "#3BB9FF",
  x_label = "Number of Samples",
  y_label = "Percentage of Proteins",
  title = NULL,
  tag = NULL,
  legend_label = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{number_of_samples}{A character string giving the column containing the number of samples}

\item{percentage}{A character string giving the column that contains the percentages}

\item{tally}{A character string giving the column that contains the number of proteins}

\item{a}{A character string giving the name of group a e.g. "Treated"}

\item{b}{A character string giving the name of group b e.g. "Control"}

\item{group}{A character string giving the column containing the group IDs}

\item{a_colour}{A character string giving the colour for group a}

\item{b_colour}{A character string giving the colour for group b}

\item{x_label}{X axis label; defaults to "Number of Samples"}

\item{y_label}{Y axis label; defaults to "Percentage of Proteins"}

\item{title}{A character string giving the plot Title}

\item{tag}{A character string giving the  plot tag}

\item{legend_label}{A character string giving the legend label}
}
\value{
Plot
}
\description{
Plots the coverage of proteins across the replicates. Requires summarised data that contains how many samples contain expression data for each protein, as a percentage of the total number of proteins.
}
