% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_overlapping.R
\name{histogram_overlapping}
\alias{histogram_overlapping}
\title{Plot two-group histograms}
\usage{
histogram_overlapping(
  df,
  x,
  a,
  b,
  group,
  colour.a = "#D55E00",
  colour.b = "#3DB7E9",
  bins = 100,
  binwidth = NULL,
  label.x = NULL,
  label.y = "Frequency",
  size.text = 20,
  title = NULL,
  tag = NULL,
  label.legend = NULL,
  alpha.bar = 0.5,
  position.legend = c(0.8, 0.2),
  colour.legend = "white",
  alpha.legend = 0.5,
  transparent = F
)
}
\arguments{
\item{df}{Dataframe containing the data to plot}

\item{x}{Character string giving the column containing the values}

\item{a}{Character string giving the name of group a e.g. "Treated"}

\item{b}{Character string giving the name of group b e.g. "Control"}

\item{group}{Character string giving the column containing the group IDs}

\item{colour.a}{Character string giving the fill color for group a; defaults to "#D55E00"}

\item{colour.b}{Character string giving the fill color for group b;defaults to "#3DB7E9"}

\item{bins}{Number of bins to use in the histogram; defaults to 100}

\item{binwidth}{Width of the bins in the histogram; defaults to NULL}

\item{label.x}{X axis label; defaults to NULL}

\item{label.y}{Y axis label; defaults to "Frequency"}

\item{size.text}{Integer value specifying the text size in the plot; defaults to 20}

\item{title}{Character string giving the plot title; defaults to NULL}

\item{tag}{Character string giving the plot tag; defaults to NULL}

\item{label.legend}{Character string giving the legend label; defaults to NULL}

\item{alpha.bar}{The transparency of the fill color; defaults to 0.5}

\item{position.legend}{A vector for the legend position; defaults to c(0.8, 0.2).}

\item{colour.legend}{Hexadecimal color code for the legend background; defaults to white.}

\item{alpha.legend}{A value between 0 and 1 for the transparency of the legend; defaults to 0.5.}

\item{transparent}{A logical value for a transparent background; defaults to False.}
}
\value{
A ggplot2 histogram plot with vertical lines indicating the median value of each group
}
\description{
This function creates a histogram plot of two groups with overlapping histograms
and adds vertical lines to indicate the median value of each group.
}
\examples{
df <- data.frame(abundances = rnorm(1000), group = sample(c("Treated", "Control"), 1000, replace = TRUE))
histogram_overlapping(df, "abundances", "Treated", "Control", "group", label.x = "Protein Abundances", label.legend = "Groups")

}
