% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_overlapping.R
\name{histogram_overlapping}
\alias{histogram_overlapping}
\title{Histogram Overlapping}
\usage{
histogram_overlapping(
  df,
  x,
  a,
  b,
  group,
  a_colour = "#F748A5",
  b_colour = "#3BB9FF",
  bins = 100,
  x_label = NULL,
  y_label = "Count",
  title = NULL,
  tag = NULL,
  legend_label = NULL,
  alpha = 0.5
)
}
\arguments{
\item{alpha}{}
}
\value{
Histogram
}
\description{
Histogram Overlapping
}
