% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_plot.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Missing data plot}
\usage{
missing_data_plot(
  df,
  sample,
  name,
  value,
  group,
  colours = c("#F748A5", "#3DB7E9"),
  x_label = "Samples",
  y_label = "Number of proteins",
  text_size = 20,
  title = NULL,
  tag = NULL,
  legend_label = NULL,
  legend_position = c(0.8, 0.2),
  legend_colour = "white",
  legend_alpha = 0.5,
  transparent = T,
  x_angle = 45,
  x_just = 1
)
}
\arguments{
\item{df}{Dataframe}

\item{sample}{A character string giving the column name for the sample IDs}

\item{name}{A character string giving the column containing present or missing}

\item{value}{A character string giving the column containing the number of proteins present or missing}

\item{group}{A character string giving the column with the group IDs e.g. "Treated" and "Control"}

\item{colours}{A character vector giving the colours; defaults to "#F748A5" and "#3DB7E9".}

\item{x_label}{X axis label; defaults to "Samples"}

\item{y_label}{Y axis label; defaults to "Number of proteins"}

\item{text_size}{Integer value specifying the text size in the plot; defaults to 20.}

\item{title}{A character string giving the plot Title}

\item{tag}{A character string giving the  plot tag}

\item{legend_label}{A character string giving the legend label}

\item{legend_position}{A vector for the legend position; defaults to c(0.8, 0.2).}

\item{legend_colour}{Hexadecimal color code for the legend background; defaults to white.}

\item{legend_alpha}{A value between 0 and 1 for the transparency of the legend; defaults to 0.5.}

\item{transparent}{A logical value for a transparent background; defaults to True.}

\item{x_angle}{A value giving the angle of the x axis text; defaults to 45.}

\item{x_just}{A value giving the horizontal justification of the x axis text; defaults to 1.}
}
\value{
plot
}
\description{
Plots the number of proteins with missing expression data for each sample. Requires a long format dataframe with the number of proteins present and missing in each sample.
}
