% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_plot.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Missing data plot}
\usage{
missing_data_plot(
  df,
  sample,
  name,
  value,
  group,
  colours = c("#F748A5", "#3DB7E9"),
  x_label = "Samples",
  y_label = "Number of proteins",
  title = NULL,
  tag = NULL,
  legend_label = NULL,
  transparent = T
)
}
\arguments{
\item{df}{Dataframe}

\item{sample}{A character string giving the column name for the sample IDs}

\item{name}{A character string giving the column containing present or missing}

\item{value}{A character string giving the column containing the number of proteins present or missing}

\item{group}{A character string giving the column with the group IDs e.g. "Treated" and "Control"}

\item{colours}{A character vector giving the colours; defaults to "#F748A5" and "#3DB7E9".}

\item{x_label}{X axis label; defaults to "Samples"}

\item{y_label}{Y axis label; defaults to "Number of proteins"}

\item{title}{A character string giving the plot Title}

\item{tag}{A character string giving the  plot tag}

\item{legend_label}{A character string giving the legend label}

\item{transparent}{A logical value for a transparent background; defaults to True.}
}
\value{
plot
}
\description{
Plots the number of proteins with missing expression data for each sample. Requires a long format dataframe with the number of proteins present and missing in each sample.
}
