% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_histogram.R
\name{pvalue_histogram}
\alias{pvalue_histogram}
\title{pvalue histogram plot for proteomic data}
\usage{
pvalue_histogram(df, x = "pvalue", text_size = 20)
}
\arguments{
\item{df}{Dataframe containing p-values}

\item{x}{A character string giving the p-value column name}

\item{text_size}{Font size; defaults to 20}
}
\value{
Histogram
}
\description{
pvalue histogram plot for proteomic data
}
