% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot for proteomic data}
\usage{
volcano_plot(
  df,
  x = "Log2fold",
  y = "Log10pvalue",
  p_value = 0.05,
  Log2foldvalue = 1
)
}
\arguments{
\item{df}{Dataframe containing Log2 fold change and Log10 p-value columns}

\item{x}{Column name (string) for Log2 fold change}

\item{y}{Column name (string) for Log10 p-value}

\item{p_value}{p=value cut off defaults to 0.05}

\item{Log2foldvalue}{Log2 fold change cut off defaults to 1}
}
\value{
Volcano plot
}
\description{
Volcano plot for proteomic data
}
\examples{
volcano_plot(example_proteomic_data)
}
