% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot for proteomic data}
\usage{
volcano_plot(
  df,
  x = "Log2fold",
  y = "Log10pvalue",
  p_value = 0.05,
  log2foldvalue = 1,
  point_size = 2,
  text_size = 20,
  axis_label_x = "Fold Change",
  axis_label_y = "p value"
)
}
\arguments{
\item{df}{Dataframe containing Log2 fold change and Log10 p value columns}

\item{x}{A character string giving the Log2 fold change column name}

\item{y}{A character string giving the Log10 p value column name}

\item{p_value}{p value cut off; defaults to 0.05}

\item{point_size}{Point size; defaults to 2}

\item{text_size}{Text size; defaults to 20}

\item{axis_label_x}{X axis label; defaults to "Fold Change"}

\item{axis_label_y}{Y axis label; defaults to "p value}

\item{Log2foldvalue}{Log2 fold change cut off; defaults to 1}
}
\value{
Volcano plot
}
\description{
Volcano plot for proteomic data
}
\examples{
volcano_plot(example_proteomic_data)
}
